refactor(entities-customization): move applier + manifest DTOs to base modules (#1895)
Layer-purity fix from /audit on branch feature/quick-wins-archi-fixes (2026-05-06).

The customization manifest applier and the descriptor delta validator are pure
domain logic (zero HTTP / FluentValidation / FusionCache imports) but lived in
Granit.Entities.Customization.Endpoints/Internal/. CLAUDE.md "Layer purity —
STRICT" requires non-HTTP domain orchestration to live in the base module.

The applier consumes the manifest value-shape DTOs (EntityManifestResponse,
EntityFormManifest, EntityProvenance, …), which themselves lived in
Granit.Entities.Endpoints/Dtos/ despite being "value-shape DTOs consumed by
domain renderers" per CLAUDE.md. Both move together — the applier cannot move
without the DTOs.

Changes
- 11 manifest DTOs: src/Granit.Entities.Endpoints/Dtos/ → src/Granit.Entities/Manifests/
  (namespace Granit.Entities.Manifests). HTTP-bound DTOs (CalendarRangeRequest,
  EntityDiscoveryItemResponse, EntityFacets, RelationAggregatesDtos) stay in
  Endpoints — they are model-bound or HTTP wire shapes.
- IManifestCustomizationApplier + NullManifestCustomizationApplier: moved to
  src/Granit.Entities/Internal/ (namespace Granit.Entities.Internal). Made the
  interface internal — InternalsVisibleTo grants Granit.Entities.Customization
  the cross-assembly visibility services.Replace<>() needs.
- EntityCustomizationManifestApplier + DescriptorDeltaValidator: moved to
  src/Granit.Entities.Customization/Internal/.
- Default no-op IManifestCustomizationApplier registration moved from
  Granit.Entities.Endpoints.AddGranitEntitiesEndpoints() →
  Granit.Entities.AddGranitEntities() (the interface lives there now).
- Replace<>() with the real applier moved from
  GranitEntitiesCustomizationEndpointsModule → GranitEntitiesCustomizationModule.
  Granit.Entities.Customization now references Granit.Entities directly.
- Tests for the applier + descriptor validator moved from
  Granit.Entities.Customization.Endpoints.Tests → Granit.Entities.Customization.Tests
  (matching where the SUTs now live).
- Removed the IManifestCustomizationApplier exclusion from
  ClassDesignTests.PublicTypesShouldNotResideInInternalNamespaces — making the
  interface internal removes the violation entirely.

Net: zero reference from Granit.Entities.Customization to
Granit.Entities.Endpoints; the manifest-applier extension point now lives
fully in the base layer, with the HTTP host only contributing its own audit
writer and route group.
