+++ src/Granit.Entities.Endpoints/EntitiesEndpoints.cs
@@ -5,6 +5,8 @@ using Granit.Entities.Endpoints.Dtos;
 using Granit.Entities.Endpoints.Internal;
 using Granit.Entities.Endpoints.Options;
 using Granit.Entities.Forms;
+using Granit.Entities.Internal;
+using Granit.Entities.Manifests;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
