+++ src/Granit.Entities.Endpoints/Extensions/EntitiesEndpointRouteBuilderExtensions.cs
@@ -68,9 +68,6 @@ public static class EntitiesEndpointRouteBuilderExtensions
         // concrete ICalendarRangeService BEFORE calling AddGranitEntitiesEndpoints,
         // or by calling Replace afterwards.
         services.TryAddSingleton<ICalendarRangeService, NullCalendarRangeService>();
-        // Layer 1 customization (ADR-053): default no-op applier. Replaced by
-        // Granit.Entities.Customization.Endpoints when that package is loaded.
-        services.TryAddScoped<IManifestCustomizationApplier, NullManifestCustomizationApplier>();
         services.AddOptions<EntitiesEndpointsOptions>();
         return services;
     }
