+++ src/Granit.Entities/Granit.Entities.csproj
@@ -10,6 +10,10 @@
     <InternalsVisibleTo Include="Granit.Entities.Tests" />
     <InternalsVisibleTo Include="Granit.Entities.Endpoints" />
     <InternalsVisibleTo Include="Granit.Entities.Endpoints.Tests" />
+    <InternalsVisibleTo Include="Granit.Entities.Customization" />
+    <InternalsVisibleTo Include="Granit.Entities.Customization.Tests" />
+    <InternalsVisibleTo Include="Granit.Entities.Customization.Endpoints" />
+    <InternalsVisibleTo Include="Granit.Entities.Customization.Endpoints.Tests" />
     <InternalsVisibleTo Include="DynamicProxyGenAssembly2" />
   </ItemGroup>
