+++ src/Granit.Entities/Manifests/EntityActivitiesManifest.cs
@@ -1,4 +1,4 @@
-namespace Granit.Entities.Endpoints.Dtos;
+namespace Granit.Entities.Manifests;
 
 /// <summary>
 /// Activities opt-in section of the per-entity manifest (ADR-046 §3). Present
